Initial commit of issue_2125 for OSX
